Refactor writeToSettings
the code in the function writeToSettings would interpret the
environment variales at the time the script was run, rather then when
the setting are sourced.  That pushes the interpretation close to the
time that they are needed so that it takes into account if the PATH or
PYTHONPATH changes from the first time this was done.  It would have
continually overwritten what the user had done before the
.dcConfig/settings was sourced.