Fix cleanup aws configs
If a user already has a set of configs that happen to have the same
name as the PROFILE then the new config is named with a -2 appended
to it before cleaning up and extra files.  Also, a note explaining
that they may have some manual work or they can just use the new one
with the new name (but it will use the specific IAM user and its
credentials).